수정: :zap: notify 아이디를 Int 로 변환

+++ android/src/main/kotlin/com/dolev/notification_when_app_is_killed/NotificationOnKillService.kt
@@ -58,7 +58,7 @@ class NotificationOnKillService: Service() {
             val notificationManager: NotificationManager =
                 getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
             notificationManager.createNotificationChannel(channel)
-            notificationManager.notify(Instant.now().epochSecond, notificationBuilder.build())
+            notificationManager.notify(Instant.now().epochSecond.toInt(), notificationBuilder.build())
         } catch (e: Exception) {
             Log.d("NotificationOnKillService", "Error showing notification", e)
         }
